# Show number of runnable/total goroutines