This commit adds a rudimentary graph for the total and runnable goroutine counts.

It only calculates the metric point once every 1<<20 nanoseconds, so about once every 1 millisecond.
Other than being a power of two so that we don't need division, this is is an arbitrary choice, I haven't checked the performance impact if the value would be different.