@martin-sucha 2024-04-24
Hi @dominikh ! Thank you for creating gotraceui, I find it really useful. I wanted to see the number of goroutine numbers, so I added an additional graph. Please let me know if you are interested to include a feature like this.

I haven't checked if it would be feasible to calculate the graph from the spans directly on every frame render for better accuracy at high zoom levels, but that is one way in which this feature could be improved.