@dominikh 2024-04-24
Thanks. Overall, the change looks really good and consistent with the surrounding code. I'd really like to get rid of the decimation to 1<<20 ns though.

This style of decimation is problematic because it is not sensitive to the shape of the data and can thus miss noteworthy events. Furthermore, there's the potential for confusion when the graph doesn't match the events exactly.

We have our own decimation for display, using the Visvalingam algorithm, so this optimization only affects memory usage and the time it takes to compute the graph, not the efficiency of displaying it. The memory usage can be quite significant: A 60 second trace of a busy Prometheus instance running on 32 cores produces 165912 + 22576840 points, and because we precompute multiple levels of decimation, this usage is multiplied by some factor. 

Comparing 1<<20 ns to 1 ns, our memory usage grows from 7 GiB to 8.8 GiB and the time to load the trace grows from 36.6s to 40.8s. We could recoup some of the time by removing all of the logic for the decimation. The memory usage is unfortunate, but we have unexplored avenues for reducing it in the future, such as compressing the data. Furthermore, a good chunk of it is due to slice growth and the GC, and we can work around that by using our `mem.BucketSlice`.

It's not feasible to calculate the graph on demand within a single frame's time. In the worst case, the user is zoomed out to view the whole range of the trace, and we can't look at all events in a busy trace within 16.6ms (and really we want to support higher refresh rates than that.) We could compute the graph asynchronously, but considering how often we have to update the graph (every time the user zooms or pans), and how expensive the computation is, this would waste a lot of CPU cycles.

How do you feel about storing all points and me addressing the memory usage in future changes?