@dominikh 2024-06-24
FTR, wanting to decimate wasn't misguided. I'm currently looking at a trace that takes 2 minutes to build the goroutine plot, and which takes 6 (out of 18 total) GB to do so. It involves a goroutine changing state every ~500 ns. Certainly something I will have to address.